requirements.txt
requests
beautifulsoup4
pdfplumber
fastapi
uvicorn
pydantic
pandas
faker
qrcode
pyotp
python-dotenv
